+++ src/main/java/io/kontur/eventapi/converter/EventDtoConverter.java
@@ -2,6 +2,7 @@ package io.kontur.eventapi.converter;
 
 import io.kontur.eventapi.entity.FeedData;
 import io.kontur.eventapi.entity.FeedEpisode;
+import io.kontur.eventapi.resource.dto.EpisodeDto;
 import io.kontur.eventapi.resource.dto.EventDto;
 import org.springframework.beans.BeanUtils;
 import org.wololo.geojson.Feature;
@@ -15,8 +16,15 @@ public class EventDtoConverter {
 
     public static EventDto convert(FeedData dataDto) {
         EventDto eventDto = new EventDto();
-        BeanUtils.copyProperties(dataDto, eventDto);
+        BeanUtils.copyProperties(dataDto, eventDto, "episodes");
+        eventDto.setEpisodes(dataDto.getEpisodes().stream().map(EventDtoConverter::convert).collect(toList()));
         return eventDto;
+    }
+
+    private static EpisodeDto convert(FeedEpisode dataDto) {
+        EpisodeDto episodeDto = new EpisodeDto();
+        BeanUtils.copyProperties(dataDto, episodeDto);
+        return episodeDto;
     }
 
     public static List<Feature> convertGeoJson(FeedData event) {