+++ src/main/java/io/kontur/eventapi/resource/EventResource.java
@@ -204,7 +204,7 @@ public class EventResource {
     @GetMapping(path = "/event", produces = {MediaType.APPLICATION_JSON_VALUE})
     @Operation(tags = "Events", summary = "returns event", description = "Returns event by its version, id and feed alias. If no version is provided the latest event version is returned.")
     @PreAuthorize("hasAuthority('read:feed:'+#feed)")
-    public ResponseEntity<FeedData> getLastEventById(@Parameter(description = "Feed name") @RequestParam(value = "feed") String feed,
+    public ResponseEntity<EventDto> getLastEventById(@Parameter(description = "Feed name") @RequestParam(value = "feed") String feed,
                                                      @Parameter(description = "Version") @RequestParam(value = "version", required = false) Long version,
                                                      @Parameter(description = "Event UUID") @RequestParam(value = "eventId") UUID eventId) {
         return eventResourceService.getEventByEventIdAndByVersionOrLast(eventId, feed, version)